pyproject.toml
[tool.poetry]
name = "action-template"
version = "1.0.0"
description = ""
authors = ["Philipp Gayret <philipp@release-engineers.com>"]
readme = "README.md"
packages = [{ include = "action_template" }]
license = "SSPL-1.0"
classifiers = ["Private :: Do not Upload"]

[tool.poetry.scripts]
"action-template" = "action_template.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
fire = "^0.5.0"
jinja2 = "^3.0.3"
pyyaml= "^6.0"
Markdown= "^3.3.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
